Add Bounded and Unbounded analytics

+++ src/main/scala/analytics/mesos/TaskPayload.scala
@@ -8,7 +8,7 @@ import io.circe.generic.semiauto.{deriveDecoder, deriveEncoder}
 
 import scala.collection.immutable.Map
 
-case class TaskPayload(data: Map[Int, SourcePair], id: Int, ops: RDataSetOp, fold: RDatasetFold, tpeA: Reified, tpeB: Reified)
+case class TaskPayload(data: Map[Int, SourcePair], id: BigInt, ops: RDataSetOp, fold: RDatasetFold, tpeA: Reified, tpeB: Reified)
 object TaskPayload {
   implicit def encoder: Encoder[TaskPayload] = deriveEncoder
   implicit def decoder: Decoder[TaskPayload] = deriveDecoder
